Add authentication with firebase and relevant views
This update includes a service that allows to connect to the firebase backend, which is initializated at boot time. Also, a couple of store modules have been defined in order to syncronize (thanks to the vuexfire package) the information between firebase and the local state.

The update also contains various placeholders for future development such as a store module for the games, a profile editor and other, currently not working, menu options.